backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pillow==8.3.2
numpy==1.21.2
requests==2.26.0
python-dotenv==0.19.0
pydantic==1.8.2
google-generativeai==0.3.1 